Move FormSync components to admin (for correct translation progress checks)
