+++ front/app/containers/Admin/projects/project/inputForm/index.tsx
@@ -3,9 +3,9 @@ import React from 'react';
 import { Box } from '@citizenlab/cl2-component-library';
 import { useParams } from 'react-router-dom';
 
+import DownloadPDFButtonWithModal from 'components/admin/FormSync/DownloadPDFButtonWithModal';
+import ExcelDownloadButton from 'components/admin/FormSync/ExcelDownloadButton';
 import { SectionTitle, SectionDescription } from 'components/admin/Section';
-import DownloadPDFButtonWithModal from 'components/FormSync/DownloadPDFButtonWithModal';
-import ExcelDownloadButton from 'components/FormSync/ExcelDownloadButton';
 import Button from 'components/UI/ButtonWithLink';
 
 import { FormattedMessage } from 'utils/cl-intl';
